Added camera pivot
I just rushed the camera last time without checking what the tutorial said. I guess it's smart to get used to attaching the camera to markers, so you can adjust the markers for what you want, then just have the camera hop between markers as needed. Not sure if I did something wrong, or if the tutorial skipped the part where they raised the camera on the y-axis.